requirements.txt
# GramGPT — зависимости
# Установка: pip install -r requirements.txt

# Telegram MTProto клиент
telethon==1.36.0

# SOCKS5/HTTP прокси
PySocks==1.7.1

# Цветной вывод
colorama==0.4.6

# Переменные окружения
python-dotenv==1.0.1

# HTTP-запросы (для AI-диалогов v0.7)
httpx==0.27.0

# ── TData импорт (опционально, нужна одна из двух) ──────────
# Вариант 1 (рекомендуется):
# pip install opentele
#
# Вариант 2:
# pip install telethon-tdata

# ── FastAPI бэкенд (v0.6, пока закомментировано) ────────────
# fastapi==0.115.0
# uvicorn==0.30.0
# sqlalchemy==2.0.35
# alembic==1.13.3
# celery==5.4.0
# redis==5.0.8